requirements.txt
# Core dependencies for domain-agnostic chatbot
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0

# AI/ML dependencies
faiss-cpu==1.11.0
huggingface-hub==0.32.6
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.3.45
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
rank-bm25==0.2.2
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
transformers==4.52.4

# Web framework (for future use)
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Utilities
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.2
mypy_extensions==1.1.0
networkx==3.5
orjson==3.10.18
packaging==24.2
Pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
zstandard==0.23.0

# Additional dependencies for document processing
python-docx==1.1.2
PyPDF2==3.0.1python-dotenv==1.0.0
